+++ clients/apple/iOS/Widgets/MobileCustomTabView.swift
@@ -3,6 +3,8 @@ import SwiftUI
 struct MobileCustomTabView<TabContent: View>: View {
     @Binding var selectedTab: TabType
     @ViewBuilder var tabContent: (TabType) -> TabContent
+    
+    @EnvironmentObject var filesModel: FilesViewModel
 
     var body: some View {
         TabView {
@@ -12,8 +14,17 @@ struct MobileCustomTabView<TabContent: View>: View {
                     systemImage: mode.systemImage
                 ) {
                     tabContent(mode)
-                }
+                }.badge(badgeCount(mode))
             }
+        }
+    }
+    
+    func badgeCount(_ mode: TabType) -> Int {
+        switch mode {
+        case .home:
+            0
+        case .sharedWithMe:
+            filesModel.pendingShares?.values.reduce(0) { $0 + $1.count } ?? 0
         }
     }
 }
